docker/requirements.txt
boto3==1.9.9
botocore==1.12.9
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
google-api-core==1.8.2
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
idna==2.8
jmespath==0.9.3
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
python-dateutil==2.7.3
pytz==2018.9
requests==2.21.0
rsa==4.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
